requirements.txt
Flask~=3.1.2
psycopg2-binary~=2.9.11
box~=0.1.5
python-box~=7.3.2